Support password expiry and 'connected' in the QEMU driver
RHEL-6 only, no upstream

The RHEL6 build of QEMU has a custom monitor command that allows
an expiry time to be set. We try to use that command first when
setting a VNC/SPICE password. If that doesn't work we fallback
to upstream way which supports this via to separate commands
(set_password and expire_password).
Also, send 'connected' attribute with command to support
action_if_connected actions.

This patch was missing check for action_if_connected being NULL
in qemuMonitorSetPasswordRH function.

Signed-off-by: Daniel Veillard <veillard@redhat.com>