daemon: Modify init script to detect upstart managed libvirtd
For https://bugzilla.redhat.com/show_bug.cgi?id=728153

This patch modifies the startup process of the libvirtd daemon. RHEL 6.2
machines use SysV init scripts for managing system processes. The VDSM
package requires using upstart script to manage libvirtd as it uses
"respawn" stanza in upstart configuration to restart libvirtd in a case
of crash.

This causes confusion as users are to manage processes with
"service" command. When libvirtd is managed using upstart and user tries
to issue commands using service command the daemon fails to start/stop
and the error messages emitted don't describe the problem helpfully.

This patch adds detection to the SysV init script, that checks if
libvirtd is managed using upstart and prints a descriptive error
message.

This is a RHEL-only patch as this problem does not occur on other
distros.

Signed-off-by: Daniel Veillard <veillard@redhat.com>