fix kvm_pv_eoi with kvmclock
https://bugzilla.redhat.com/show_bug.cgi?id=860971

When both kvmclock and kvm_pv_eoi are configured (either disabled or
enabled) libvirt will generate invalid CPU specification due to the
fact that even though kvmclock causes the CPU to be specified, it
doesn't set have_cpu flag to true (and the new kvm_pv_eoi as well).
This patch fixes the issue and adds a test exactly for that to show
that it is fixed correctly (and also to keep it that way in the future
of course).

(cherry picked from commit 5d692cc7143084d841de79888fc249040ff40061)