util: extend virGetUserID and virGetGroupID to support names and IDs
https://bugzilla.redhat.com/show_bug.cgi?id=860519

This patch updates virGetUserID and virGetGroupID to be able to parse a
user or group name in a similar way to coreutils' chown. This means that
a numeric value with a leading plus sign is always parsed as an ID,
otherwise the functions try to parse the input first as a user or group
name and if this fails they try to parse it as an ID.

This patch includes Peter Krempa's changes to correctly handle errors
returned by getpwnam_r and getgrnam_r.
(cherry picked from commit 0b237296ef8a702df238d97fb9ab2534c8820330)