doc: update description about user/group in qemu.conf
https://bugzilla.redhat.com/show_bug.cgi?id=860519

As a side effect of changes in the functions virGetUserID and
virGetGroupID, the user and group configurations for DAC in qemu.conf
are now able to accept both names and IDs, supporting a leading plus
sign to ensure that a numeric value will not be interpreted as a name.

This patch updates the comments in qemu.conf, including a description of
this new behavior.
(cherry picked from commit 60dea2c6bff0d3ee0df3bda9c8521bdfc42407f3)