Revert "Use XDG Base Directories instead of storing in home directory"
https://bugzilla.redhat.com/show_bug.cgi?id=859331

This reverts commit 32a9aac2e04c991340b66c855a1095e4e6445e54.

The reason for this are selinux policies that can't be modified in
RHEL6 to suit these needs, thus this commit is causing problems on
RHEL6 systems.

Many commits between this and 32a9aa were fixing some minor
documentation issues, some of them are needed to be kept and so on.
Due to these commits, there is a lot of conflicts, so sorry for not
explaining them all.

RHEL-only.

Conflicts:
    AUTHORS -- I've kept the person in the file
    daemon/libvirtd-config.c
    daemon/libvirtd.c
    src/libvirt.c
    src/network/bridge_driver.c
    src/nwfilter/nwfilter_driver.c
    src/qemu/qemu_driver.c
    src/remote/remote_driver.c
    src/secret/secret_driver.c
    src/storage/storage_driver.c
    src/uml/uml_driver.c
    src/util/util.c
    src/util/util.h
    src/util/virauth.c
    tools/virsh.c -- conflict with virsh-splitting series