Add support for SUSPEND_DISK event
https://bugzilla.redhat.com/show_bug.cgi?id=839661

This patch adds support for SUSPEND_DISK event; both lifecycle and
separated.  The support is added for QEMU, machines are changed to
PMSUSPENDED, but as QEMU sends SHUTDOWN afterwards, the state changes
to shut-off.  This and much more needs to be done in order for libvirt
to work with transient devices, wake-ups etc.  This patch is not
aiming for that functionality.
(cherry picked from commit 7ba5defb5a34bb82de2f16467c2b8d157cd14d2d)