storage: make it easier to find file within chain
https://bugzilla.redhat.com/show_bug.cgi?id=772088

In order to temporarily label files read/write during a commit
operation, we need to crawl the backing chain and find the absolute
file name that needs labeling in the first place, as well as the
name of the file that owns the backing file.

* src/util/storage_file.c (virStorageFileChainLookup): New
function.
* src/util/storage_file.h: Declare it.
* src/libvirt_private.syms (storage_file.h): Export it.
(cherry picked from commit 5eaf6054473cf0a1f4753e5e43280f89ec91c4d1)