blockjob: implement shallow commit flag in qemu
https://bugzilla.redhat.com/show_bug.cgi?id=772088

Now that we can crawl the chain of backing files, we can do
argument validation and implement the 'shallow' flag.  In
testing this, I discovered that it can be handy to pass the
shallow flag and an explicit base, as a means of validating
that the base is indeed the file we expected.

* src/qemu/qemu_driver.c (qemuDomainBlockCommit): Crawl through
chain to implement shallow flag.
* src/libvirt.c (virDomainBlockCommit): Relax API.
(cherry picked from commit 0a220e22251fbb047791f50fd2d15b49f274fc07)